refactor: model apis
